requirements.txt
aiogram>=3.14.0
aiofiles>=23.0.0
python-dotenv>=1.0.0
